chore: Update dependencies and enhance Python bindings
- Updated Rust version in Cargo.toml to 1.70 for improved compatibility.
- Enhanced clippy_warning_counter.py to better track warning occurrences by package.
- Added gzip decompression support in examples/Cargo.toml for dataset handling.
- Improved documentation across various modules, ensuring clarity and completeness.
- Refined error handling in the autograd and nn modules for better stability.
- Updated README and SPRINT_MS37_CONTEXT.md to reflect recent changes and future directions.

+++ nn/src/nas/evaluator.rs
@@ -92,7 +92,7 @@ impl ArchitectureEvaluator for SimpleEvaluator {
         metrics.insert("complexity_score".to_string(), num_layers);
         metrics.insert(
             "efficiency_ratio".to_string(),
-            accuracy / (num_params.max(1.0) as f64),
+            accuracy / num_params.max(1.0),
         );
 
         Ok(EvaluationResult {
